+++ RadRefinements/Patches/CratePatches.cs
@@ -12,8 +12,15 @@ namespace RadRefinements.Patches
             [HarmonyPatch("UpdateLookText")]
             public static void AddNumItemsText(ShipItemCrate __instance, Good ___goodC, CrateInventory ___crateInventory)
             {
-                if (!__instance.sold || __instance.amount > 0 || !(bool)___crateInventory || (bool)___goodC && ___goodC.GetMissionIndex() > -1)
+                if (!Plugin.enableCrateItemDescription.Value ||
+                    !__instance.sold ||
+                    __instance.amount > 0 ||
+                    !(bool)___crateInventory ||
+                    ((bool)___goodC &&
+                    ___goodC.GetMissionIndex() > -1))
+                {
                     return;
+                }                    
 
                 __instance.lookText = $"{__instance.lookText}\n{___crateInventory.containedItems.Count()} items";
             }